テキストチャンネルのUIを共通化
