+++ app/routes/channels+/_text/model/profile.ts
